Developer Dashboard, Become/Update Developer, Any List/Show Published TrainingModules

+++ src/main/java/acme/form/DeveloperDashboard.java
@@ -15,7 +15,7 @@ public class DeveloperDashboard extends AbstractForm {
 	private Integer				numTrainingSessionsLink;
 	private Double				avgTrainingModulesTime;
 	private Double				devTrainingModulesTime;
-	private Double				minTrainingModulesTime;
-	private Double				maxTrainingModulesTime;
+	private Integer				minTrainingModulesTime;
+	private Integer				maxTrainingModulesTime;
 
 }
